Allow singular pattern for basic word searching
This doesn't yet support using a single pattern with alternations
instead of multiple patterns, it simply allows for searching the
word list for words matching a pattern.

This makes the superpuzzition command useful other than for finding
superpuzzitions, so I may want to rename it to find-words or similar.

This is mainly a precursor to supporting searches involving grids with
cells with more than two letters superimposed. It was simpler to
implement variable length word lists first before introducing the
complexity of combining multiple similarity scores.